refactor: extract perl-error and perl-tokenizer microcrates
- Extracted error handling logic to  crate.
- Extracted tokenizer logic to  crate.
- Consolidated  into  crate.
- Updated  to use the new microcrates.
- Fixed circular dependencies and updated tests.
