+++ Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "crates/perl-lsp",
     "crates/perl-dap",
     "xtask", "crates/perl-ast", "crates/perl-token", "crates/perl-quote", "crates/perl-pragma", "crates/perl-edit", "crates/perl-builtins", "crates/perl-regex", "crates/perl-heredoc",
+    "crates/perl-error", "crates/perl-tokenizer",
 ]
 exclude = [
     "tree-sitter-perl",
@@ -47,6 +48,8 @@ perl-pragma = { path = "crates/perl-pragma", version = "0.8.8" }
 perl-lexer = { path = "crates/perl-lexer", version = "0.8.8" }
 perl-ast = { path = "crates/perl-ast", version = "0.8.8" }
 perl-heredoc = { path = "crates/perl-heredoc", version = "0.0.1" }
+perl-error = { path = "crates/perl-error", version = "0.8.8" }
+perl-tokenizer = { path = "crates/perl-tokenizer", version = "0.8.8" }
 perl-position-tracking = { path = "crates/perl-position-tracking", version = "0.8.8" }
 perl-symbol-types = { path = "crates/perl-symbol-types", version = "0.8.8" }
 perl-symbol-table = { path = "crates/perl-symbol-table", version = "0.8.8" }
