+++ crates/perl-parser-core/Cargo.toml
@@ -33,6 +33,8 @@ perl-edit = { workspace = true }
 perl-builtins = { workspace = true }
 perl-regex = { workspace = true }
 perl-heredoc = { workspace = true }
+perl-error = { workspace = true }
+perl-tokenizer = { workspace = true }
 thiserror = "2.0.18"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
