+++ crates/perl-parser-core/src/engine/mod.rs
@@ -3,7 +3,6 @@
 /// Abstract Syntax Tree (AST) definitions for Perl parsing.
 pub use perl_ast::ast;
 /// Experimental second-generation AST (work in progress).
-#[allow(missing_docs)]
 pub use perl_ast::v2 as ast_v2;
 /// Edit tracking for incremental parsing.
 pub use perl_edit as edit;
@@ -23,4 +22,4 @@ pub use perl_pragma as pragma_tracker;
 pub use perl_quote as quote_parser;
 /// Parser utilities and helpers.
 pub use perl_regex as regex_validator;
-pub mod util;
+
