+++ crates/perl-parser-core/src/engine/position/mod.rs
@@ -2,8 +2,8 @@
 //!
 //! This module re-exports position types from `perl-position-tracking`.
 
-pub use perl_position_tracking::{Position, Range};
+pub use perl_position_tracking::{Position, Range, LineIndex};
 pub use perl_position_tracking::{offset_to_utf16_line_col, utf16_line_col_to_offset};
 pub use perl_position_tracking::mapper::{LineEnding, PositionMapper};
 
-pub mod line_index;
+
