--- crates/perl-parser-core/src/tokens/mod.rs
@@ -1,6 +1,0 @@
-//! Token stream and trivia utilities for the parser.
-
-pub mod token_stream;
-pub mod token_wrapper;
-pub mod trivia;
-pub mod trivia_parser;
