+++ crates/perl-tokenizer/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "perl-tokenizer"
+version = "0.8.8"
+edition.workspace = true
+rust-version.workspace = true
+authors = ["Tree-sitter Perl Contributors"]
+description = "Token stream and utilities for Perl parser"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/EffortlessMetrics/tree-sitter-perl-rs"
+
+[dependencies]
+perl-lexer = { workspace = true }
+perl-token = { workspace = true }
+perl-error = { workspace = true }
+perl-position-tracking = { workspace = true }
+perl-ast = { workspace = true }
+
+[dev-dependencies]
+
+[lints]
+workspace = true
