+++ crates/perl-tokenizer/src/lib.rs
@@ -1,0 +1,14 @@
+//! Token stream and trivia utilities for the parser.
+
+pub mod token_stream;
+pub mod token_wrapper;
+pub mod trivia;
+pub mod trivia_parser;
+pub mod util;
+
+pub use token_stream::TokenStream;
+pub use perl_token::{Token, TokenKind};
+pub use token_wrapper::TokenWithPosition;
+pub use trivia::{Trivia, TriviaToken};
+pub use trivia_parser::{TriviaParserContext, TriviaPreservingParser};
+
