+++ crates/perl-tokenizer/src/util.rs
